+++ MobSoft/app/src/main/java/com/example/mobsoft/mobsoft/ui/persondetails/PersonDetailsPresenter.java
@@ -52,7 +52,7 @@ public class PersonDetailsPresenter extends Presenter<PersonDetailsScreen> {
         super.detachScreen();
     }
 
-    public void getPerson(final int id) {
+    public void getPerson(final long id) {
         executor.execute(new Runnable() {
             @Override
             public void run() {
